[bndtools launch] Unify launch delegates' pre-launch validation logic
Fixes #3954.
